refactor(pdf): extract region line analysis
